Moved tag component to tag directory for better structure
